+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+##[v0.2.0]
+
+- Add [#10] profiling tool for typical verification of names from OCRed texts.
+- Add [#9] partial matches for names that did not match fully.
+- Add [#8] optimization/parallelizaton of matching algorithms.
 
 ## [v0.1.0]
 
@@ -18,9 +23,12 @@
 
 This document follows [changelog guidelines]
 
-[v0.1.0]: https://github.com/gnames/gnmatcher/compare/v0.1.0...v0.1.0
+[v0.2.0]: https://github.com/gnames/gnmatcher/compare/v0.1.0...v0.2.0
+[v0.1.0]: https://github.com/gnames/gnmatcher/compare/v0.0.0...v0.1.0
 [v0.0.0]: https://github.com/gnames/gnmatcher/tree/v0.0.0
 
+[#10]: https://github.com/gnames/gnmatcher/issues/10
+[#9]: https://github.com/gnames/gnmatcher/issues/9
 [#8]: https://github.com/gnames/gnmatcher/issues/8
 [#7]: https://github.com/gnames/gnmatcher/issues/7
 [#6]: https://github.com/gnames/gnmatcher/issues/6